feat: :alien: Replace axios with ky in API calls
This commit replaces the usage of axios with ky in API calls to improve the performance and simplify the code. Ky is a smaller and faster library than axios, and it has a simpler API that is easier to use. This change also removes the withCredentials option, as ky does not support it by default, but it can be manually added if needed.